requirements.txt
matplotlib==3.8.2
pandas==2.2.0
python_docx==1.1.0